Move extension interfaces to extension module
* move IDraggable to extensions.../drag/*
* move ISwipeable to extensions.../swipe/*
* move IExtendedDraggable to extensions.../drag/*
