+++ library-extensions/src/main/java/com/mikepenz/fastadapter_extensions/drag/IExtendedDraggable.java
@@ -1,10 +1,10 @@
-package com.mikepenz.fastadapter_extensions;
+package com.mikepenz.fastadapter_extensions.drag;
 
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
 import android.view.View;
 
-import com.mikepenz.fastadapter.IDraggable;
+import com.mikepenz.fastadapter_extensions.drag.IDraggable;
 import com.mikepenz.fastadapter.IItem;
 
 /**
